requirements.txt
Cerberus==1.1
lxml==3.7.3
PyYAML==3.12
requests==2.13.0
